Update and rename ej1-2.py to ej3.py
Ejercicio 3
